refactor: fix pr ppl homepage
