+++ src/block/home_campaign_demo/home_campaign_demo.module.css
@@ -7,7 +7,7 @@
 
 .title_section_one {
     font-family: "Aladin", sans-serif;
-    @apply text-4xl md:text-7xl font-bold text-center mb-12;
+    @apply text-7xl font-bold text-center mb-12;
 }
 
 .container_section_one {
@@ -15,7 +15,7 @@
 }
 
 .grid_section_one {
-    @apply grid grid-cols-1 md:grid-cols-2 gap-8 items-center;
+    @apply grid grid-cols-2 gap-8 items-center;
 }
 
 .image_container {
