+++ src/block/presentation/presentation.module.css
@@ -1,6 +1,6 @@
 .text_presentation {
     font-family: "Aladin", sans-serif;
-    @apply text-[40px];
+    @apply text-4xl;
 }
 
 .presentation {
